requirements.txt
requests
pycryptodome
curl_cffi
aiohttp
certifi
browser_cookie3
websockets
js2py
quickjs
flask
flask-cors
python-telegram-bot
beautifulsoup4
configparser
newsapi-python